requirements.txt
requests
python-jose[cryptography]
fastapi
passlib[bcrypt]
uvicorn
python-multipart
pytest
pygame
sqlalchemy
